First pass at configuration.
First pass at a blind configuration. This code assumes we have a fresh unit. It doesn't attempt to recover from any problems. If the unit doesn't accept the command we don't know that and blindly carry on with the next step. This should work on fresh units, but won't be able to handle a reattempt at a configuration.